case of % i

+++ 01-printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mfaoussi <mfaoussi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/24 08:11:32 by mfaoussi          #+#    #+#             */
-/*   Updated: 2024/01/01 10:01:18 by mfaoussi         ###   ########.fr       */
+/*   Updated: 2024/01/01 10:05:33 by mfaoussi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,16 +21,16 @@
 # include <limits.h>
 
 int				ft_printf(const char *format, ...);
-char			*hex_converter(unsigned int nb, int mode);
-int				hextcount(unsigned int nb);
-void			hexwrite(unsigned int nb, int mode, int *count);
+char			*hex_converter(unsigned long long nb, int mode);
+int				hextcount(unsigned long long nb);
+void			hexwrite(unsigned long long nb, int mode, int *count);
 void			ft_putchar_fd(char c, int fd, int *count);
 void			ft_putnbr_fd(int n, int fd, int *count);
 void			ft_putstr_fd(char *s, int fd, int *count);
 char			*ft_strdup(const char *str);
 void			print_null(char *s, int fd, int *count);
-unsigned int	checker(unsigned int nb);
-void			print_pointer(long address, int *count);
+unsigned int	checker(unsigned long long nb);
+void			print_pointer(unsigned long long address, int *count);
 
 // void	ft_putchar_fd(char c, int fd);
 // void	ft_putstr_fd(char *s, int fd);
